requirements.txt
asgiref==3.5.2
Django==4.0.4
django-extensions==3.1.5
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1